+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/evm/config/EvmConfiguration.java
@@ -20,8 +20,6 @@ import com.github.benmanes.caffeine.cache.Caffeine;
 import com.hedera.mirror.web3.evm.contracts.execution.MirrorEvmMessageCallProcessor;
 import com.hedera.mirror.web3.evm.contracts.execution.MirrorEvmMessageCallProcessorV30;
 import com.hedera.mirror.web3.evm.contracts.operations.HederaBlockHashOperation;
-import com.hedera.mirror.web3.evm.contracts.operations.HederaSelfDestructOperation;
-import com.hedera.mirror.web3.evm.contracts.operations.HederaSelfDestructOperationV038;
 import com.hedera.mirror.web3.evm.properties.MirrorNodeEvmProperties;
 import com.hedera.mirror.web3.evm.store.contract.EntityAddressSequencer;
 import com.hedera.mirror.web3.repository.properties.CacheProperties;
@@ -37,6 +35,8 @@ import com.hedera.node.app.service.evm.contracts.operations.HederaExtCodeHashOpe
 import com.hedera.node.app.service.evm.contracts.operations.HederaExtCodeSizeOperation;
 import com.hedera.services.contracts.gascalculator.GasCalculatorHederaV22;
 import com.hedera.services.evm.contracts.operations.HederaPrngSeedOperation;
+import com.hedera.services.evm.contracts.operations.HederaSelfDestructOperation;
+import com.hedera.services.evm.contracts.operations.HederaSelfDestructOperationV038;
 import com.hedera.services.txns.crypto.AbstractAutoCreationLogic;
 import com.hedera.services.txns.util.PrngLogic;
 import java.util.HashMap;
@@ -45,6 +45,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.function.BiPredicate;
+import java.util.function.Predicate;
 import javax.inject.Provider;
 import lombok.RequiredArgsConstructor;
 import org.hyperledger.besu.datatypes.Address;
@@ -96,11 +97,11 @@ public class EvmConfiguration {
     private final MirrorNodeEvmProperties mirrorNodeEvmProperties;
     private final GasCalculatorHederaV22 gasCalculator;
     private final HederaBlockHashOperation hederaBlockHashOperation;
-    private final HederaSelfDestructOperation hederaSelfDestructOperation;
-    private final HederaSelfDestructOperationV038 hederaSelfDestructOperationV038;
     private final AbstractAutoCreationLogic autoCreationLogic;
     private final EntityAddressSequencer entityAddressSequencer;
     private final PrecompiledContractProvider precompilesHolder;
+    private final BiPredicate<Address, MessageFrame> addressValidator;
+    private final Predicate<Address> systemAccountDetector;
 
     @Bean(CACHE_MANAGER_CONTRACT_STATE)
     CacheManager cacheManagerState() {
@@ -214,7 +215,9 @@ public class EvmConfiguration {
     }
 
     @Bean
-    EVM evm030(final HederaPrngSeedOperation prngSeedOperation) {
+    EVM evm030(
+            final HederaPrngSeedOperation prngSeedOperation,
+            final HederaSelfDestructOperation hederaSelfDestructOperation) {
         return evm(
                 gasCalculator,
                 mirrorNodeEvmProperties,
@@ -226,7 +229,9 @@ public class EvmConfiguration {
     }
 
     @Bean
-    EVM evm034(final HederaPrngSeedOperation prngSeedOperation) {
+    EVM evm034(
+            final HederaPrngSeedOperation prngSeedOperation,
+            final HederaSelfDestructOperation hederaSelfDestructOperation) {
         return evm(
                 gasCalculator,
                 mirrorNodeEvmProperties,
@@ -238,7 +243,9 @@ public class EvmConfiguration {
     }
 
     @Bean
-    EVM evm038(final HederaPrngSeedOperation prngSeedOperation) {
+    EVM evm038(
+            final HederaPrngSeedOperation prngSeedOperation,
+            final HederaSelfDestructOperationV038 hederaSelfDestructOperationV038) {
         return evm(
                 gasCalculator,
                 mirrorNodeEvmProperties,
@@ -252,6 +259,16 @@ public class EvmConfiguration {
     @Bean
     HederaPrngSeedOperation hederaPrngSeedOperation(final GasCalculator gasCalculator, final PrngLogic prngLogic) {
         return new HederaPrngSeedOperation(gasCalculator, prngLogic);
+    }
+
+    @Bean
+    HederaSelfDestructOperation hederaSelfDestructOperation(final GasCalculator gasCalculator) {
+        return new HederaSelfDestructOperation(gasCalculator, addressValidator);
+    }
+
+    @Bean
+    HederaSelfDestructOperationV038 hederaSelfDestructOperationV038(final GasCalculator gasCalculator) {
+        return new HederaSelfDestructOperationV038(gasCalculator, addressValidator, systemAccountDetector);
     }
 
     @Bean
